% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagst.R
\name{roastDir}
\alias{roastDir}
\title{ROAST directed test}
\usage{
roastDir(array, design, inds)
}
\arguments{
\item{array}{expression array}

\item{design}{design matrix (intercept included by default)}

\item{inds}{indices of transcripts to be tested}
}
\description{
ROAST directed test
}
\examples{
roastDir(arr,des,ind)
}
