% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagst.R
\name{roastMix}
\alias{roastMix}
\title{ROAST undirected/mixed test}
\usage{
roastMix(array, design, inds, samp_size = 2000)
}
\arguments{
\item{array}{expression array. Genes along rows.}

\item{design}{design matrix (intercept included by default)}

\item{inds}{indices of transcripts to be tested}

\item{samp_size}{number of samples off of which empirical p-value is calculated (not permutation based, but sampling from parametric distribution)}
}
\description{
ROAST undirected/mixed test
}
\examples{
arr <- matrix(rnorm(100*100),nrow=100,ncol=100)
des <- rep(0:1,each=50)
ind <- sample(100,10)
roastMix(arr,des,ind)
}
