% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagst.R
\name{JAGSTself}
\alias{JAGSTself}
\title{JAGST self-contained test}
\usage{
JAGSTself(array, design, inds)
}
\arguments{
\item{array}{expression array.  Genes along rows.}

\item{design}{design matrix (intercept included by default)}

\item{inds}{indices of transcripts to be tested}
}
\description{
JAGST self-contained test
}
\examples{
JAGSTself(arr,des,ind)
}
