% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagst.R
\name{JAGSTself}
\alias{JAGSTself}
\title{JAGST self-contained test}
\usage{
JAGSTself(array, design, inds)
}
\arguments{
\item{array}{expression array. Genes along rows.}

\item{design}{design matrix (intercept included by default)}

\item{inds}{indices of transcripts to be tested}
}
\description{
JAGST self-contained test
}
\examples{
dat <- matrix(rnorm(100*100),nrow=100,ncol=100)
design <- rep(0:1,each=50)
inds <- sample(100,10)
JAGSTself(arr,des,ind)
}
