% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagst.R
\name{roastDir}
\alias{roastDir}
\title{ROAST directed test}
\usage{
roastDir(array, design, inds)
}
\arguments{
\item{array}{expression array. Genes along rows.}

\item{design}{design matrix (intercept included by default)}

\item{inds}{indices of transcripts to be tested}
}
\description{
ROAST directed test
}
\examples{
arr <- matrix(rnorm(100*100),nrow=100,ncol=100)
des <- rep(0:1,each=50)
ind <- sample(100,10)
roastDir(arr,des,ind)
}
