% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagst.R
\name{roastMix}
\alias{roastMix}
\title{ROAST undirected/mixed test}
\usage{
roastMix(array, design, inds, samp_size = 2000)
}
\arguments{
\item{array}{expression array}

\item{design}{design matrix (intercept included by default)}

\item{inds}{indices of transcripts to be tested}

\item{samp_size}{number of samples off of which empirical p-value is calculated (not permutation based, but sampling from parametric distribution)}
}
\description{
ROAST undirected/mixed test
}
\examples{
roastMix(arr,des,ind,samp_size)
}
